move explosion static method to utils
